Restructured and reorganized methods. Added addUserToEvent.
Restructured and reorganized the methods and what they do. Moved some of
the code in AndroidIO methods that interacts with the database
(especially the queries) into the DBIO directory. The result of this is
a more modular approach to interaction between all of the PHP methods,
allowing for more and easier interaction.

Added the addUserToEvent method along with about 4 other methods it
depends on (created for modularity) as well as a JSON configuration
file. This event checks every possible detail related to an event, the
inviter, and the invitee to determine whether or not an invitee should
be added to an event, and do so.