Added a new function and updated an existing one.
Added a new function getImagesByPrivacyAndPurpose, which returns all of
the images from the database table that are of the specified privacy and
purpose. These images are in the form of Image arrays which contain
data. 

Modified the getEventDataByMember function to return not only the events
to which the specified user belongs but also the events of which the
specified user is a host.

Updated the old getFriends function to also return the username.
Added the getUserSentFriendRequestUsers to return a list of users to
whom friend requests were sent.