Added the getEventDataByTopNLikes and getEventDataByTopNDislikes functions to the EventRequest and Event class.
The first function (+ db function) gets the data of an entire event for
all of the events that have the top N count of likes, where N is the
input value.
This second function (+ db function) gets the data of an entire event
for all of the events that have the top N count of dislikes, where N is
the input value.