Added the updateEvent and getImageProfileMaxAmount classes and methods.
Added the updateEvent class and method which accepts all of the event's
properties (except the counts) as input parameters. It then updates the
properties of the event that has the input event identifier with the
provided non-null properties. "Null"s that were provided instead of
values in the corresponding property's position will result in that
property not being updated for the event.
Added the getImageProfileMaxAmount class and method which returns the
integer that represents the maximum amount of profile images allowed per
user.