Fixed rejectFriend & created cancelFriend.
Fixed the FriendshipStatusRequest/rejectFriend function to work
correctly. The input parameters are uid1 [first parameter] (identifier
for the user performing the reject action) and uid2 [second parameter]
(identifier for the user on whom the reject action is performed). If a
row of data exists in the database representing uid2 having sent a
friend request to uid1 previously, this will return a success message,
otherwise a fail message.
Created the FriendshipStatusRequest/cancelFriend function. The input
parameters are uid1 [first parameter] (identifier for the user
performing the cancel action) and uid2 [second parameter] (identifier
for the user on whom the cancel action is performed). If a row of data
exists in the database representing uid1 having sent a friend request to
uid2 previously, this will return a success message, otherwise a fail
message.