Updated the ServerRequestMethods.getUsers function.
Updated the ServerRequestMethods.getUsers function. Now, the UID of the
user performing the search is input as the first parameter, whereas the
search string is input as the second parameter. The result will be
candidate users in addition to their friendship status regarding the
searchedByUid user. 
ADDITIONALLY, the getFriendStatus PHP file was converted to a
getFriendStatus.fetchFriendshipStatus function with uid_1 and uid_2 as
input parameters, and the FriendshipStatusRequest.getFriendshipStatus
method was added to handle the server request call. In other words,
AndroidIO/FriendshipStatusRequest/getFriendshipStatus replaces the
previous Older/getFriendStatus.