Added the getEventDataByRadius function.
Added the (db/)getEventDataByRadius function to the EventRequest file in
the AndroidIO & DBIO packages. If the longitude, latitude, and radius
are provided, this method will return a JSON String which contains an
array of Events and all of its associated data as well as the distance,
in miles,  from the specified latitude and longitude.