Added db/getLiveEvent functions.
Added db/getLiveEvent functions, which get the data of the events that
the existing getEventData functions returned, but with one exception:
only events that are within the boundary that is at most one day before
the event's start date and at most one day after the event's end date.
Added getEventDataByTopNRatings functions, which gets the data of the
most rated N events, where the rating is the ratio of likes to the sum
of likes and dislikes.

Made minor bugfixes.