Created Event + User + Image related files/functions.
[Java/PHP] Created Event + User + Image related files/functions, which
are AndroidIO/UserImageRequest.php - addImagesToEvent() and
AndroidIO/UserImageRequest.php - getImagesByEid & DBIO/UserImage.php -
fetchImagesByEid. The first adds images that were previously uploaded
into the database to the Event with the specified eid. The second
returns all of the images and their data that have been added to the
Event with the specified eid. 

[Java/PHP] Fixed the NullPointerException error in the
AndroidIO/UserImageRequest.php - getImagesByUidAndPurpose() &
DBIO/UserImage.php - fetchImagesByUidAndPurpose() as well as
AndroidIO/UserImageRequest.php - getImagesByPrivacyAndPurpose() &
DBIO/UserImage.php - getImagesByPrivacyAndPurpose() methods, caused by
the extra "eid" parameter, which is no longer returned because it is no
longer a part of the image object. 

[PHP] Optimized the uploadImage function.