Created the unFriend and getEventUsersData classes/methods.
Created the FriendshipStatusRequest.php - unFriend function, which
unfriends a friend for a user, both of whose UIDs have been specified
and exist as a friendship in the database.
Created the EventUserRequest.php - getEventUsersData function, which
returns all of the user and event-user related data of all of the users
that have joined/left/been-removed-from the specified event.