Added the UserComment(Request) files.
Added the UserComment(Request) files, which contain the
(db)setObjectComment and (db)getObjectComments functions. 

setObjectComment adds the specified comment of the specified user to the
specified object type with the specified object id at the specified
timestamp. If the comment is a reply to an existing comment, its parent
UCID (parent user comment ID) must also have been specified for this
request. If this is an existing comment, only its contents and timestamp
will be updated instead.

getObjectComments retrieves the comments and all of their associated
data for the specified type of object with the specified identifier.
This may, for example, retrieve all of the comments that were added to a
particular event, user, or user image.