Added and modified files and functions.
Added the <db>setEuiEventProfileSequence function to the
EventUserImage<Request> files, which sets the sequence of the event user
image in the event profile. 
Modified the UserImage<Request> [get/fetch]ImagesByEid functions to
allow for another input parameter. euiProfileIndicator, or Event User
Image Profile Indicator narrows down the result set to event user images
that belong to a profile (if input is set to "true"), event user images
that do not belong to a profile (if input is set to "false"), or all
event user images (if input is set to "null").
Fixed the name of the DBIO/addImageToEvent function to
dbAddImageToEvent.