Added the [db]setEventUser function.
Added the [db]setEventUser function to the EventUser[Request] file,
which updates the user with input uid of the event with input eid with
the input values. The input values that are to update the current values
may be actual values or NULL. Whether or not they are updated is
determined by the setOrNot Boolean array, true representing that the
value with the provided key should be updated. 

Added the dbGetEventUserEncoded function to the EventUser file, which
gets the latest Event User data while keeping the reference data
encoded. This is used as part of the [db]setEventUser function to
compare the current Event User values against the new Event User values. 

Added the dbGetEventUserInviteStatusTypeCode function to the
ReferenceData file, which gets the Event User Invite Status Type Code
for the input Event User Invite Status Type Label. This is used as part
of the [db]setEventUser function to encode the Event User Invite Status
Type Label provided by the user and update it for the Event User data
value.