Added, modified, and removed files and functions.
Added the UserImage[Request]/[db]getImagesByUid function, which gets all
of the images for the specified uid, optionally filtered on whether they
are profile images or non-profile images. This function replaced
UserImage[Request]/[db]getImagesByUidAndPurpose, as the "AndPurpose"
part has been incorporated into the
UserImage[Request]/[db]getImagesByUid through the userProfileIndicator
parameter. 
Modified the UserImage[Request]/[db]getImagesByEid function, to be as
similar to [db]getImagesByUid as possible. It now also returns the
euiEventProfileSequence, euiInsertTimestamp, and euiUpdateTimestamp.
(eui = Event User Image). 
Removed the UserImage[Request]/[db]getImagesByPrivacyAndPurpose
function. It was determined to be impractical, and is no needed in the
presence of the two functions above. 

Fixed the NewsFeed/dbGetNewsFriendCreatedEvent function to return
profile images for the friend, not for the event host. 
Fixed the EventUser/dbGetEventUserData method to now return the eid
(event user identifier) and uid (user identifier) in the correct order
in the query.