Added the User<Request>/<db>getUserProfileData file/function.
Added the UserRequest/getUserProfileData file/function, which gets the
profile data for the user with the specified uid. In other words, gets
anything about the user that includes the username, first name, last
name, profile image and its thumbnail, count of friends, a few friends,
a few events, etc. This retrieves more data than getUserData.

Added support functions to be called on by the above file/function:
- dbGetCountFriends, which gets the count of friends that the user with
the specified UID has
- dbGetCountHostedEvents, which gets the count of events that the user
with the specified UID hosted 
- dbGetCountJoinedEvents, which gets the count of events that the user
with the specified UID joined 
- dbGetCountImages, which gets the count of images that the user with
the specified UID uploaded 
- dbGetEventDataByTopNRandom, which gets the top N random events that
the user with the specified UID is a part of 
- dbGetFriendsByTopNRandom, which gets the top N users that the user
with the specified UID is friends with 
- dbGetImagesLatestNByUid, which gets the latest N images that the user
with the specified UID uploaded